Only one map created, markers of search results are added to the map until Reset button pressed which will delete all the markers in markersArray and empty the search results (checkboxes).
Reload button will reload all the markers in the markersArray onto the map.

Organisation ID search follows the same way as name search.

Once the search result has been added to the map, the correspondent checkboxs are disabled.

Problem of timeslider: how to deal with the start point and end point? Originally, the two points are fixed with the events of one organisation. How about now if mutiple organisation involved?